moved the footer map from index.jsx, removed the console on about us
