+++ client/src/index.jsx
@@ -1,6 +1,5 @@
-import React, {useState, useEffect} from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import axios from 'axios';
 import './index.css';
 
 import {
@@ -17,16 +16,6 @@ import {
 } from './components';
 
 function App() {
-  const [footerInfo, setFooterInfo] = useState([]);
-  const LoadFooter = async() => {
-      const answer = await axios.get('http://localhost:3001/api/footers');
-      setFooterInfo(answer.data);
-    };
-    
-    useEffect(() => {
-        LoadFooter();
-    },[]);
-  
   return (
     <section className="app">
       <meta name="Description" content="One page Aqua Jr."/>
@@ -41,24 +30,9 @@ function App() {
       <Testimony />
       <CallToBlog />
       <Contact />
-      <div className="FooterComponent">
-        { footerInfo.map ((FooterComponent) =>{
-          return(
-            <Footer
-                Address = { FooterComponent.Address }
-                Phone = { FooterComponent.Phone}
-                Email = {FooterComponent.Email }
-                Facebook = { FooterComponent.Facebook }
-                Instagram = { FooterComponent.Instagram }
-                LinkedIn = { FooterComponent.LinkedIn }
-            />
-          );
-        })}
-      </div>
- 
-      
+      <Footer/>
     </section>
   );
 }
-
+ 
 ReactDOM.render(<App />, document.getElementById('root'))
